Refactor report and hits.js
